Colocate Api helpers with their request methods.
