--- src/api/Api.helpers.cpp
@@ -1,19 +1,0 @@
-#include <vector>
-
-#include "api/Api.hpp"
-#include "api/types.hpp"
-
-bool Api::IsLoggedIn()
-{
-    return !login.member_id.empty();
-}
-
-std::vector<PortfolioPositionsResponse::EventPosition> Api::GetEventPositions()
-{
-    return positions.event_positions;
-}
-
-std::vector<PortfolioPositionsResponse::MarketPosition> Api::GetMarketPositions()
-{
-    return positions.market_positions;
-}
